+++ prisma/migrations/20230706155702_add_quote_table/migration.sql
@@ -1,0 +1,7 @@
+-- CreateTable
+CREATE TABLE "quote" (
+    "id" SERIAL NOT NULL,
+    "content" TEXT NOT NULL,
+
+    CONSTRAINT "quote_pkey" PRIMARY KEY ("id")
+);